+++ UniversalDeviceToolkit.CLI.Lib/IpcRequest.cs
@@ -29,4 +29,6 @@ public class IpcRequest
     public string? Name { get; init; }
 
     public string? Value { get; init; }
+
+    public string? AuthToken { get; set; }
 }