+++ UniversalDeviceToolkit.Lib/Controllers/Sensors/FpsSensorController.cs
@@ -13,10 +13,10 @@ namespace LenovoLegionToolkit.Lib.Controllers.Sensors
 {
     public class FpsSensorController : IDisposable
     {
-        [DllImport("user32.dll")]
+        [DllImport("user32.dll", CharSet = CharSet.Unicode, CallingConvention = CallingConvention.StdCall)]
         private static extern IntPtr GetForegroundWindow();
 
-        [DllImport("user32.dll")]
+        [DllImport("user32.dll", CharSet = CharSet.Unicode, CallingConvention = CallingConvention.StdCall)]
         private static extern uint GetWindowThreadProcessId(IntPtr hWnd, out uint lpdwProcessId);
 
         public class FpsData